requirements.txt
python-telegram-bot==13.15
schedule==1.2.1
python-dotenv==1.0.1